requirements.txt

streamlit
faker
pandas
python-stdnum
